+++ UPR/Projekt/vars.h
@@ -14,7 +14,7 @@ bool loadTextures(WindowProperties *windowProperties, SDL_Renderer *renderer);
 void freeTextures(WindowProperties *windowProperties);
 
 // load window properties
-bool loadWindowProperties(WindowProperties *windowProperties, SDL_Renderer *renderer, Colors *colors, int SCALE, SDL_Window *window);
+bool loadWindowProperties(WindowProperties *windowProperties, SDL_Renderer *renderer, Colors *colors, float SCALE, SDL_Window *window);
 
 // free window properties
 void freeWindowProperties(WindowProperties *windowProperties);
